Refactor library export metadata module
